Documents/GitHub/pulser/tools/claude_behavior_checklist.py
#!/usr/bin/env python3
from pathlib import Path
import yaml
import os

# Define the checklist as a structured dictionary
claude_behavior_checklist = {
    "agent": "claude",
    "category": "Behavioral Parity Validation",
    "checklist": {
        "Identity & Capability Parity": [
            "Claude is formally recognized as a Pulser agent in the registry",
            "Claude's capabilities mirror Claude Code CLI functions",
            "Claude's responses include agent-like metadata"
        ],
        "Context + Memory Fidelity": [
            "Claude recalls user-defined preferences from prior sessions",
            "Claude matches Pulser's YAML-based persistent memory setup",
            "Claude auto-syncs with Claudia logs",
            "Claude respects `pulser-status`, `claudia-log`, and session trust contexts"
        ],
        "UI/UX Behavior Match": [
            "Claude mirrors Pulser CLI behaviors",
            "Claude CLI supports aliases identical to Pulser's",
            "Claude respects the structured natural prompt format"
        ],
        "Security & Permission": [
            "Claude enforces token limits, sandbox scope, and agent boundary awareness",
            "Claude prompts for escalation or task handoff when exceeding role"
        ],
        "Update & Version Integrity": [
            "Claude is monitored for API deprecation or model updates",
            "Claude version parity is confirmed with Claude Code changelog",
            "Pulser auto-logs changes in Claude behavior"
        ],
        "Testing & Behavioral Regression": [
            "Stress tests simulate load and prompt ambiguity",
            "Regression tests ensure Claude returns identical responses for same task",
            "A/B trust scoring is recorded"
        ],
        "Design Philosophy Compliance": [
            "Claude follows clarity, consistency, and utility principles",
            "Emotional intelligence detection for escalation"
        ],
        "Metrics Logging": [
            "Session consistency score",
            "Factual coherence rating",
            "Completion latency and retry rate",
            "Honesty rate via 'I don't know'",
            "Context window accuracy %"
        ],
        "Sync Validation Script": [
            "CLI command `pulser-validate-claude-sync` runs behavioral parity checks"
        ]
    }
}

def main():
    # Create output directory if it doesn't exist
    config_dir = os.path.expanduser("~/.pulser/config")
    os.makedirs(config_dir, exist_ok=True)
    
    # Save to file
    file_path = Path(f"{config_dir}/claude_behavior_checklist.yaml")
    with open(file_path, "w") as file:
        yaml.dump(claude_behavior_checklist, file, sort_keys=False)
    
    print(f"Claude behavior checklist saved to: {file_path}")
    return file_path.name

if __name__ == "__main__":
    main()