Documents/GitHub/pulser/requirements.txt
openai>=1.2.0
anthropic>=0.3.11
faiss-cpu>=1.7.4
numpy>=1.23.0
pandas>=1.5.3
scikit-learn>=1.2.2
python-dotenv>=1.0.0
tqdm>=4.66.1
PyYAML>=6.0
click>=8.1.3
rich>=13.7.0
pytest>=7.4.3
matplotlib>=3.7.2